requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0

# ML/AI frameworks
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# LLM frameworks
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Vector database
qdrant-client>=1.7.0

# Graph database
neo4j>=5.14.0

# Multimodal processing
Pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
opencv-python>=4.8.0
moviepy>=1.0.3
openai-whisper>=20231117

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.10.0
python-magic-bin>=0.4.14; platform_system == "Windows"

# Audio processing
librosa>=0.10.0
pydub>=0.25.1

# Evaluation
deepeval>=0.20.90

# Agent frameworks
autogen-agentchat>=0.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tenacity>=8.2.0
tqdm>=4.66.0
aiohttp>=3.9.0
httpx>=0.25.0

# UI
streamlit>=1.28.0
